requirements.txt
Django==1.8.7
django-haystack==2.4.1
django-taggit==0.17.6
wheel==0.24.0
Whoosh==2.7.0
